feat(ctl): optimize version detection and logging
Reduces unnecessary overhead when detecting the BeeGFS version for a particular volume by skipping
the check if a particular version of CTL is installed and just running the list nodes command. The
logic is the same regardless if the check fails for command not found or a failure executing a
command due to a mismatched version or other external reason so there is no reason to differentiate.

Improved logging by dropping unnecessary stack traces and updating chwrap to return the standard
error and return code when a CTL binary is not installed (instead of a scary panic).